Moved result handling for image choose and capture intent to FileWidget
