+++ collect_app/src/main/java/org/odk/collect/android/widgets/FileWidget.java
@@ -3,9 +3,14 @@ package org.odk.collect.android.widgets;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
+import android.support.v4.content.FileProvider;
 
 import org.javarosa.form.api.FormEntryPrompt;
+import org.odk.collect.android.BuildConfig;
+import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.utilities.ApplicationConstants;
+import org.odk.collect.android.utilities.FileUtils;
+import org.odk.collect.android.utilities.ImageConverter;
 import org.odk.collect.android.utilities.MediaUtils;
 import org.odk.collect.android.widgets.interfaces.BinaryWidget;
 
@@ -49,6 +54,49 @@ public abstract class FileWidget extends QuestionWidget implements BinaryWidget 
             case ApplicationConstants.RequestCodes.VIDEO_CHOOSER:
                 getWidgetAnswerListener().saveChosenFile(data.getData());
                 break;
+            case ApplicationConstants.RequestCodes.DRAW_IMAGE:
+            case ApplicationConstants.RequestCodes.ANNOTATE_IMAGE:
+            case ApplicationConstants.RequestCodes.SIGNATURE_CAPTURE:
+            case ApplicationConstants.RequestCodes.IMAGE_CAPTURE:
+                saveCapturedImage();
+                break;
+            case ApplicationConstants.RequestCodes.IMAGE_CHOOSER:
+                getWidgetAnswerListener().saveChosenImage(data.getData());
+                break;
         }
+    }
+
+
+    /*
+     * We saved the image to the tempfile_path, but we really want it to
+     * be in: /sdcard/odk/instances/[current instnace]/something.jpg so
+     * we move it there before inserting it into the content provider.
+     * Once the android image capture bug gets fixed, (read, we move on
+     * from Android 1.6) we want to handle images the audio and video
+     */
+    protected void saveCapturedImage() {
+        // The intent is empty, but we know we saved the image to the temp file
+        ImageConverter.execute(Collect.TMPFILE_PATH, this, getContext());
+        File fi = new File(Collect.TMPFILE_PATH);
+
+        // Revoke permissions granted to this file due its possible usage in the camera app
+        Uri uri = FileProvider.getUriForFile(getContext(), BuildConfig.APPLICATION_ID + ".provider", fi);
+        FileUtils.revokeFileReadWritePermission(getContext(), uri);
+        saveImageToInstanceFolder(fi);
+        getWidgetAnswerListener().saveAnswersForCurrentScreen(false);
+    }
+
+    protected void saveImageToInstanceFolder(File imageFile) {
+        String instanceFolder = getFormController().getInstanceFile().getParent();
+        String s = instanceFolder + File.separator + System.currentTimeMillis() + ".jpg";
+
+        File nf = new File(s);
+        if (!imageFile.renameTo(nf)) {
+            Timber.e("Failed to rename %s", imageFile.getAbsolutePath());
+        } else {
+            Timber.i("Renamed %s to %s", imageFile.getAbsolutePath(), nf.getAbsolutePath());
+        }
+
+        setBinaryData(nf);
     }
 }
